requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
unstructured
libmagic
python-magic
# python-magic-bin (this is windows specific)
qdrant-client
fastembed
python-dotenv
httpx